close mongo connection after execution
