+++ src/utils/index.ts
@@ -1,8 +1,8 @@
-import { areMongoIdEqual, buildMongo, buildObjectId, mongoIdType } from './buildMongo';
+import { areMongoIdEqual, buildObjectId, mongoIdType, buildRunMongo } from './mongo';
 import { dateBuilder } from './dateBuilder';
 import { dependencyManager } from './dependencyManager';
 import { buildHandlingErrorController } from './express';
 
-export { areMongoIdEqual, buildHandlingErrorController, buildMongo, buildObjectId, dateBuilder, dependencyManager };
+export { areMongoIdEqual, buildHandlingErrorController, buildObjectId, buildRunMongo, dateBuilder, dependencyManager };
 
 export type { mongoIdType };
